scrapper/prt_scrapper/requirements.txt
werkzeug==2.2.2
flask==2.2.2
flask-httpauth==4.7.0
flask-sqlalchemy==3.0.2
flask-caching==2.0.2
psycopg2-binary==2.9.5
python-dotenv==0.21.1
selenium==4.8.0
requests==2.28.2